Migrate PolygonShape to Polygon
This was absolutely choking my PC when I connected to a server (140fps to 40fps), dottrace wasn't super helpful but at least we can just use the fixedarray polys instead of the array ones. Mostly wanted it around the collisionmanager area but figured we just start porting stuff over to the new version.